introduce CacheResponseStrategy manager service
